requirements.txt
fastapi
uvicorn[standard] 
pillow 
torch 
torchvision 
facenet-pytorch 
scikit-learn 
joblib 
scikit-image 
opencv-python 
python-multipart
